core/telegram_interface/requirements-fixed.txt
# Core Telegram Bot (let it manage its own httpx version)
python-telegram-bot==20.7

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
cryptography==41.0.7
schedule==1.2.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Voice processing and AI (essential only)
openai==1.3.0
